Updated parameters to include namespace
I absolutely loved the original script, but the only thing that didn't work for me was that all my deployments are in custom namespaces. 
I have updated the way parameters are read in so they are no longer positional and now support an optional namespace (which defaults to 'default'). 
This change also allows for other non-positional arguments to be supplied to the bash script if the functionality is ever extended.